+++ apps/explorer/lib/explorer/chain/cache/gas_price_oracle.ex
@@ -10,6 +10,8 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
       from: 2
     ]
 
+  alias EthereumJSONRPC.Blocks
+
   alias Explorer.Chain.{
     Block,
     Wei
@@ -25,8 +27,17 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
     ttl_check_interval: :timer.seconds(1),
     callback: &async_task_on_deletion(&1)
 
+  @doc """
+  Get `safelow`, `average` and `fast` percentile of transactions gas prices among the last `num_of_blocks` blocks
+  """
+  @spec get_average_gas_price(pos_integer(), pos_integer(), pos_integer(), pos_integer()) ::
+          {:error, any} | {:ok, %{String.t() => nil | float, String.t() => nil | float, String.t() => nil | float}}
   def get_average_gas_price(num_of_blocks, safelow_percentile, average_percentile, fast_percentile) do
-    latest_gas_price_query =
+    safelow_percentile_fraction = safelow_percentile / 100
+    average_percentile_fraction = average_percentile / 100
+    fast_percentile_fraction = fast_percentile / 100
+
+    fee_query =
       from(
         block in Block,
         left_join: transaction in assoc(block, :transactions),
@@ -35,32 +46,116 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
         where: transaction.gas_price > ^0,
         group_by: block.number,
         order_by: [desc: block.number],
-        select: min(transaction.gas_price),
+        select: %{
+          slow_gas_price:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^safelow_percentile_fraction,
+              transaction.gas_price
+            ),
+          average_gas_price:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^average_percentile_fraction,
+              transaction.gas_price
+            ),
+          fast_gas_price:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^fast_percentile_fraction,
+              transaction.gas_price
+            ),
+          slow:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^safelow_percentile_fraction,
+              transaction.max_priority_fee_per_gas
+            ),
+          average:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^average_percentile_fraction,
+              transaction.max_priority_fee_per_gas
+            ),
+          fast:
+            fragment(
+              "percentile_disc(?) within group ( order by ? )",
+              ^fast_percentile_fraction,
+              transaction.max_priority_fee_per_gas
+            )
+        },
         limit: ^num_of_blocks
       )
 
-    latest_gas_prices =
-      latest_gas_price_query
-      |> Repo.all(timeout: :infinity)
-
-    latest_ordered_gas_prices =
-      latest_gas_prices
-      |> Enum.map(fn %Wei{value: gas_price} -> Decimal.to_integer(gas_price) end)
-
-    safelow_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, safelow_percentile)
-    average_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, average_percentile)
-    fast_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, fast_percentile)
-
-    gas_prices = %{
-      "slow" => safelow_gas_price,
-      "average" => average_gas_price,
-      "fast" => fast_gas_price
-    }
+    gas_prices = fee_query |> Repo.all(timeout: :infinity) |> process_fee_data_from_db()
 
     {:ok, gas_prices}
   catch
     error ->
       {:error, error}
+  end
+
+  defp process_fee_data_from_db([]) do
+    %{
+      "slow" => nil,
+      "average" => nil,
+      "fast" => nil
+    }
+  end
+
+  defp process_fee_data_from_db(fees) do
+    fees_length = Enum.count(fees)
+
+    %{
+      slow_gas_price: slow_gas_price,
+      average_gas_price: average_gas_price,
+      fast_gas_price: fast_gas_price,
+      slow: slow,
+      average: average,
+      fast: fast
+    } =
+      fees
+      |> Enum.reduce(
+        &Map.merge(&1, &2, fn
+          _, v1, v2 when nil not in [v1, v2] -> Decimal.add(v1, v2)
+          _, v1, v2 -> v1 || v2
+        end)
+      )
+      |> Map.new(fn
+        {key, nil} -> {key, nil}
+        {key, value} -> {key, Decimal.div(value, fees_length)}
+      end)
+
+    json_rpc_named_arguments = Application.get_env(:explorer, :json_rpc_named_arguments)
+
+    {slow_fee, average_fee, fast_fee} =
+      case {nil not in [slow, average, fast], EthereumJSONRPC.fetch_block_by_tag("pending", json_rpc_named_arguments)} do
+        {true, {:ok, %Blocks{blocks_params: [%{base_fee_per_gas: base_fee}]}}} when not is_nil(base_fee) ->
+          base_fee_wei = base_fee |> Decimal.new() |> Wei.from(:wei)
+
+          {
+            priority_with_base_fee(slow, base_fee_wei),
+            priority_with_base_fee(average, base_fee_wei),
+            priority_with_base_fee(fast, base_fee_wei)
+          }
+
+        _ ->
+          {gas_price(slow_gas_price), gas_price(average_gas_price), gas_price(fast_gas_price)}
+      end
+
+    %{
+      "slow" => slow_fee,
+      "average" => average_fee,
+      "fast" => fast_fee
+    }
+  end
+
+  defp priority_with_base_fee(priority, base_fee) do
+    priority |> Wei.from(:wei) |> Wei.sum(base_fee) |> Wei.to(:gwei) |> Decimal.to_float() |> Float.ceil(2)
+  end
+
+  defp gas_price(value) do
+    value |> Wei.from(:wei) |> Wei.to(:gwei) |> Decimal.to_float() |> Float.ceil(2)
   end
 
   defp num_of_blocks, do: Application.get_env(:explorer, __MODULE__)[:num_of_blocks]
@@ -100,40 +195,6 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
       end)
 
     {:update, task}
-  end
-
-  defp gas_price_percentile_to_gwei(gas_prices, percentile) do
-    gas_price_wei = percentile(gas_prices, percentile)
-
-    if gas_price_wei do
-      gas_price_gwei = Wei.to(%Wei{value: Decimal.from_float(gas_price_wei)}, :gwei)
-
-      gas_price_gwei_float = gas_price_gwei |> Decimal.to_float()
-
-      if gas_price_gwei_float > 0.01 do
-        gas_price_gwei_float
-        |> Float.ceil(2)
-      else
-        gas_price_gwei_float
-      end
-    else
-      nil
-    end
-  end
-
-  @spec percentile(list, number) :: number | nil
-  defp percentile([], _), do: nil
-  defp percentile([x], _), do: x
-  defp percentile(list, 0), do: Enum.min(list)
-  defp percentile(list, 100), do: Enum.max(list)
-
-  defp percentile(list, n) when is_list(list) and is_number(n) do
-    s = Enum.sort(list)
-    r = n / 100.0 * (length(list) - 1)
-    f = :erlang.trunc(r)
-    lower = Enum.at(s, f)
-    upper = Enum.at(s, f + 1)
-    lower + (upper - lower) * (r - f)
   end
 
   # By setting this as a `callback` an async task will be started each time the